+++ scripts/test/Signature.t
@@ -12,14 +12,20 @@ use utf8;
 
 use vars (qw($Self));
 
-use Kernel::System::ObjectManager;
+# get helper object
+$Kernel::OM->ObjectParamAdd(
+    'Kernel::System::UnitTest::Helper' => {
+        RestoreDatabase            => 1,
+    },
+);
+my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 
-# get needed objects
+# get signature object
 my $SignatureObject = $Kernel::OM->Get('Kernel::System::Signature');
 
 # add signature
-my $SignatureNameRand0 = 'example-signature' . int( rand(1000000) );
-my $Signature          = "Your OTRS-Team
+my $SignatureName = 'signature' . $Helper->GetRandomID();
+my $SignatureText = "Your OTRS-Team
 
 <OTRS_CURRENT_UserFirstname> <OTRS_CURRENT_UserLastname>
 
@@ -30,8 +36,8 @@ Email: hot\@florida.com - Web: http://hot.florida.com/
 --";
 
 my $SignatureID = $SignatureObject->SignatureAdd(
-    Name        => $SignatureNameRand0,
-    Text        => $Signature,
+    Name        => $SignatureName,
+    Text        => $SignatureText,
     ContentType => 'text/plain; charset=iso-8859-1',
     Comment     => 'some comment',
     ValidID     => 1,
@@ -47,12 +53,12 @@ my %Signature = $SignatureObject->SignatureGet( ID => $SignatureID );
 
 $Self->Is(
     $Signature{Name} || '',
-    $SignatureNameRand0,
+    $SignatureName,
     'SignatureGet() - Name',
 );
 $Self->True(
-    $Signature{Text} eq $Signature,
-    'SignatureGet() - Signature',
+    $Signature{Text} eq $SignatureText,
+    'SignatureGet() - Signature text',
 );
 $Self->Is(
     $Signature{ContentType} || '',
@@ -71,21 +77,17 @@ $Self->Is(
 );
 
 my %SignatureList = $SignatureObject->SignatureList( Valid => 0 );
-my $Hit = 0;
-for ( sort keys %SignatureList ) {
-    if ( $_ eq $SignatureID ) {
-        $Hit = 1;
-    }
-}
 $Self->True(
-    $Hit eq 1,
-    'SignatureList()',
+    exists $SignatureList{$SignatureID} && $SignatureList{$SignatureID} eq $SignatureName,
+    "SignatureList() contains the signature $SignatureName",
 );
 
-my $SignatureUpdate = $SignatureObject->SignatureUpdate(
+my $SignatureNameUpdate = $SignatureName . ' - Update';
+my $SignatureTextUpdate = $SignatureText . ' - Update';
+my $SignatureUpdate     = $SignatureObject->SignatureUpdate(
     ID          => $SignatureID,
-    Name        => $SignatureNameRand0 . '1',
-    Text        => $Signature . '1',
+    Name        => $SignatureNameUpdate,
+    Text        => $SignatureTextUpdate,
     ContentType => 'text/plain; charset=utf-8',
     Comment     => 'some comment 1',
     ValidID     => 2,
@@ -101,11 +103,11 @@ $Self->True(
 
 $Self->Is(
     $Signature{Name} || '',
-    $SignatureNameRand0 . '1',
+    $SignatureNameUpdate,
     'SignatureGet() - Name',
 );
 $Self->True(
-    $Signature{Text} eq $Signature . '1',
+    $Signature{Text} eq $SignatureTextUpdate,
     'SignatureGet() - Signature',
 );
 $Self->Is(
@@ -123,5 +125,13 @@ $Self->Is(
     2,
     'SignatureGet() - ValidID',
 );
+
+%SignatureList = $SignatureObject->SignatureList( Valid => 1 );
+$Self->False(
+    exists $SignatureList{$SignatureID},
+    "SignatureList() does not contain invalid signature $SignatureNameUpdate",
+);
+
+# cleanup is done by RestoreDatabase
 
 1;