+++ scripts/test/StdAttachment.t
@@ -15,16 +15,24 @@ use vars (qw($Self));
 use Kernel::System::VariableCheck qw(:all);
 
 # get needed objects
-my $ConfigObject           = $Kernel::OM->Get('Kernel::Config');
-my $HelperObject           = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 my $MainObject             = $Kernel::OM->Get('Kernel::System::Main');
 my $StdAttachmentObject    = $Kernel::OM->Get('Kernel::System::StdAttachment');
 my $StandardTemplateObject = $Kernel::OM->Get('Kernel::System::StandardTemplate');
 
+# get helper object
+$Kernel::OM->ObjectParamAdd(
+    'Kernel::System::UnitTest::Helper' => {
+        RestoreDatabase => 1,
+    },
+);
+my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
+
+my $Home = $Kernel::OM->Get('Kernel::Config')->Get('Home');
+my $Location;
+
 # file checks
 for my $File (qw(xls txt doc png pdf)) {
-    my $Location = $ConfigObject->Get('Home')
-        . "/scripts/test/sample/StdAttachment/StdAttachment-Test1.$File";
+    $Location = $Home . "/scripts/test/sample/StdAttachment/StdAttachment-Test1.$File";
 
     my $ContentRef = $MainObject->FileRead(
         Location => $Location,
@@ -160,11 +168,10 @@ for my $File (qw(xls txt doc png pdf)) {
 
 # attachment -> templates tests
 my $UserID   = 1;
-my $RandomID = $HelperObject->GetRandomID();
+my $RandomID = $Helper->GetRandomID();
 
 # create a new attachment
-my $Location = $ConfigObject->Get('Home')
-    . "/scripts/test/sample/StdAttachment/StdAttachment-Test1.png";
+$Location = $Home . "/scripts/test/sample/StdAttachment/StdAttachment-Test1.png";
 
 my $ContentRef = $MainObject->FileRead(
     Location => $Location,
@@ -173,12 +180,14 @@ my $ContentRef = $MainObject->FileRead(
 
 my $Content = ${$ContentRef};
 
-my $AttachmentID = $StdAttachmentObject->StdAttachmentAdd(
-    Name        => 'Standard Attachment' . $RandomID,
+my $AttachmentName = 'Standard Attachment' . $RandomID;
+my $Filename       = 'StdAttachment-Test1.xml';
+my $AttachmentID   = $StdAttachmentObject->StdAttachmentAdd(
+    Name        => $AttachmentName,
     ValidID     => 1,
     Content     => $Content,
     ContentType => 'text/xml',
-    Filename    => 'StdAttachment-Test1.xml',
+    Filename    => $Filename,
     Comment     => 'Some Comment',
     UserID      => $UserID,
 );
@@ -368,7 +377,7 @@ $Self->True(
             UserID             => $UserID,
         },
         ExpectedResults => {
-            $AttachmentID => 'Standard Attachment' . $RandomID,
+            $AttachmentID => $AttachmentName,
         },
         Success => 1,
     },
@@ -399,33 +408,33 @@ for my $Test (@Tests) {
     }
 }
 
-# cleanup
-
-# delete relation
-$Success = $StdAttachmentObject->StdAttachmentStandardTemplateMemberAdd(
-    AttachmentID       => $AttachmentID,
-    StandardTemplateID => $TemplateID,
-    Active             => 0,
-    UserID             => $UserID
+my %StdAttachmentList = $StdAttachmentObject->StdAttachmentList(
+    Valid => 0,
 );
 $Self->True(
-    $Success,
-    "StdAttachmentStandardTemplateMemberAdd() removal for Attachment -> Template tests | with True",
+    exists $StdAttachmentList{$AttachmentID} && $StdAttachmentList{$AttachmentID} eq "$AttachmentName ( $Filename )",
+    "StdAttachmentList() contains the StdAttachment $AttachmentName",
 );
 
-$Success = $StdAttachmentObject->StdAttachmentDelete( ID => $AttachmentID );
+my $Update = $StdAttachmentObject->StdAttachmentUpdate(
+    ID      => $AttachmentID,
+    Name    => $AttachmentName,
+    ValidID => 2,
+    UserID  => $UserID,
+);
 $Self->True(
-    $Success,
-    "StdAttachemntDelete() for Attachment -> Template tests | with True",
+    $Update,
+    'StdAttachmentUpdate() - StdAttachment is set to invalid',
 );
 
-$Success = $StandardTemplateObject->StandardTemplateDelete(
-    ID => $TemplateID,
+%StdAttachmentList = $StdAttachmentObject->StdAttachmentList(
+    Valid => 1,
+);
+$Self->False(
+    exists $StdAttachmentList{$AttachmentID},
+    "StdAttachmentList() does not contain the StdAttachment $AttachmentName",
 );
 
-$Self->True(
-    $Success,
-    "StandardTemplateDelete() for Attachment -> Template tests | with True",
-);
+# cleanup is done by RestoreDatabase
 
 1;